docs: document the control-size ramp under spacing, not layout
controlSize sizes controls — heights, paddings, icon boxes — which is a
sizing concern, not a layout one. It only sat on the Layout foundations
page because it is breakpoint-responsive. Move its showcase to the
Spacing page, beside the space scale, where a reader looking for control
dimensions would expect it.

Extracts the section into ControlSizeShowcase and folds it into the
Spacing page's intro; the Layout page keeps breakpoints, content width,
layers, and aspect ratios.
