+++ apps/web/src/app/[locale]/design-system/foundations/spacing/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { DocPage } from "#src/components/design-system/doc-page.tsx";
+import { ControlSizeShowcase } from "#src/components/design-system/sections/tokens/control-size-showcase.tsx";
 import { SpaceScaleShowcase } from "#src/components/design-system/sections/tokens/space-scale-showcase.tsx";
 import { t } from "#src/i18n.ts";
 import type { PageProps } from "#src/types.ts";
@@ -20,11 +21,12 @@ export default function SpacingPage() {
     <DocPage
       title={t({ en: "Spacing", zh: "间距" })}
       description={t({
-        en: "One rem-based scale of eighteen steps, each drawn to true size as a ruler.",
-        zh: "一套以 rem 为基准、共十八级的阶梯，皆按真实尺寸以标尺呈现。",
+        en: "The rem-based spacing scale — eighteen steps drawn to true size as a ruler — and the responsive control-size ramp that sets the dimensions of every interactive control.",
+        zh: "以 rem 为基准的间距阶梯——十八级按真实尺寸如标尺般呈现——以及决定每个交互控件尺寸的响应式控件尺寸阶梯。",
       })}
     >
       <SpaceScaleShowcase />
+      <ControlSizeShowcase />
     </DocPage>
   );
 }
